telegram_bot/requirements.txt
aiogram
langchain-ollama
langchain-core
mcp-use
python-dotenv
mcp[cli]
httpx